requirements.txt
# Function dependencies, for example:
# package>=version
Flask==2.0.3
google-api-core==2.10.1
google-auth==2.11.1
google-auth-oauthlib==0.5.3
google-cloud-bigquery==3.3.2
google-cloud-bigquery-storage==2.16.0
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mysql-connector-python==8.0.30
numpy==1.23.3
lxml==4.9.1
oauthlib==3.2.1
packaging==21.3
pandas==1.5.0
pandas-gbq==0.17.9
pyarrow==9.0.0
pyorc==0.8.0
PyMySQL==1.0.2
python-dateutil==2.8.2
requests==2.28.1
requests-oauthlib==1.3.1
SQLAlchemy==1.4.41
fsspec==2022.10.0
gcsfs==2022.10.0
fastavro==1.7.0
openpyxl==3.0.10
xlrd==2.0.1
xmltodict==0.13.0
